+++ app/Http/Controllers/CollegeController.php
@@ -91,6 +91,11 @@ class CollegeController extends Controller
     public function deleteCollege(Request $request)
     {
         Gate::authorize('deleteCollege', CollegeController::class);
+        if (ValidateHelper::validateData($request, [
+            'id' => 'required|integer'
+        ])) {
+            return  ResponseHelper::clientError();
+        }
         try {
             $college = College::findOrFail($request->id);
             $college->delete();
@@ -133,6 +138,11 @@ class CollegeController extends Controller
     public function retrieveCollege(Request $request)
     {
         Gate::authorize('retrieveCollege', CollegeController::class);
+        if (ValidateHelper::validateData($request, [
+            'id' => 'required|integer'
+        ])) {
+            return  ResponseHelper::clientError();
+        }
         $attributes = ['arabic_name', 'english_name', 'phone', 'email', 'description', 'youtube', 'x_platform', 'facebook', 'telegram', 'logo_url'];
         $conditionAttribute = ['id' => $request->id];
         try {
@@ -147,6 +157,11 @@ class CollegeController extends Controller
     public function retrieveEditableCollege(Request $request)
     {
         Gate::authorize('retrieveEditableCollege', CollegeController::class);
+        if (ValidateHelper::validateData($request, [
+            'id' => 'required|integer'
+        ])) {
+            return  ResponseHelper::clientError();
+        }
         $attributes = ['arabic_name', 'english_name', 'phone', 'email', 'description', 'youtube', 'x_platform', 'facebook', 'telegram', 'logo_url'];
         $conditionAttribute = ['id' => $request->id];
         try {
